Flatten instance, domains, reports; delete Hunter.Api.HTTPClient
